proxy: Guard against servers that do not send an 'error' field
Avoid errors when 'error' is not present in the server's response.
The specification says that it "must be null if there was no error",
but not all servers implement the specification correctly.

Assume that a missing 'error' field means no-errors.
The example http://www.raboof.com/projects/jayrock/demo.ashx
service is one such server that does not return an error
on success but otherwise returns correct results.