servicehandler: refactor and simplify the API
The original API required users to use the servicemethod decorator.
This is now optional.

The original API expected an object to be passed in as the
ServiceHandler's underlying service.  We now allow a dict to be passed
in as well.

This can be used with the new servicemodule() function which scans a
module for methods and returns a dictionary of the public (non-
underscore) methods.

Signed-off-by: David Aguilar <davvid@gmail.com>