wsgi: use WsgiContentReader to read data from wsgi.input
When reading from the wsgi.input file object, a size must be
specified to avoid blocking.

If the length of the wsgi.imput content is larger than the max
chunk size then the content will be read in chunks up to the
number of bytes specified by the Content-Length header.

Signed-off-by: Neil Barber <neil.barber@disneyanimation.com>